Restrict the version bound on data-default-class and wrapped.
According to the PVP specification, packages providing orphan instances
must depend on the minor version of the packages providing the class and
the instance type.  In this case it seems obvious that neither `wrapped`
nor `data-default-class` will provide this instance in future versions,
but it may as well conform to the specification.